% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{add_input_pop}
\alias{add_input_pop}
\title{Shortcut for adding tooltip to a shiny input}
\usage{
add_input_pop(tag, popover_text, placement = "top")
}
\arguments{
\item{tag}{The object to add the tooltip to}

\item{popover_text}{Text to display}

\item{placement}{Where to display}
}
\value{
Input, now with popover
}
\description{
This can be piped onto the end of a Shiny input. See examples in UI functions.
}
