% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{App server function}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny inputs}

\item{output}{Shiny outputs}

\item{session}{SHiny session}
}
\description{
This is called by \code{run_app()}.
}
