% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{box_pop_title}
\alias{box_pop_title}
\title{Add help icon to box title}
\usage{
box_pop_title(title, popover_text, placement = "top", px_from_right = 10)
}
\arguments{
\item{title}{The box title text}

\item{popover_text}{Text to display in the popover}

\item{placement}{Where to show the popover text}

\item{px_from_right}{Position from the right side of the box. This will need to
be adjusted depending on the other content in the box, e.g. if there is a collapsible
icon, sidebar icon, etc. See examples in app.}
}
\value{
HTML to be inserted in box title
}
\description{
Shortcut for adding an info icon to a box title. This is used in the "title"
argument of a Shinydashboard box.
}
