% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{cache_admin2_geometry}
\alias{cache_admin2_geometry}
\title{Retrieve and store geometry for countries}
\usage{
cache_admin2_geometry(ISO3s, overwrite = FALSE)
}
\arguments{
\item{ISO3s}{Character vector of ISO3s to get maps for}

\item{overwrite}{if TRUE overwrites any existing files}
}
\description{
Queries the UNHCR API to get Admin2 layer for the specified countries, then
simplifies the geometry and stores in inst/geom. To enable quick retrieval
of maps. This is intended to just be run occasionally.
}
