% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{colour_input}
\alias{colour_input}
\title{Text input for map colour}
\usage{
colour_input(id, initial_colour)
}
\arguments{
\item{id}{ID for slider}

\item{initial_colour}{Initial colour value}
}
\value{
HTML code for slider
}
\description{
Thin wrapper for textInput().
}
