% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Country names, codes and flags}
\format{
A data frame with 249 rows and 4 columns
}
\source{
countrycodes package, plus github for flags
}
\usage{
country_codes
}
\description{
Data frame with country names, ISO codes (alpha 2 and alpha 3), and links to
svg flags (used e.g. in dropdown selectors).
}
\keyword{datasets}
