% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcome_tour.R
\name{create_guide}
\alias{create_guide}
\title{Interactive tour}
\usage{
create_guide(id)
}
\arguments{
\item{id}{The namespace - not sure if actually used.}
}
\value{
Tour via cicerone
}
\description{
Generates an interactive tour using the cicerone package. The "el" entries
correspond to IDs defined in the UI. Sometimes I have had to wrap things
to get an ID. Not perfect but does the job.
}
\details{
This is called in the "Welcome" tab. See also the cicerone documentation to
understand how this works.
}
