% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_analysis.R
\name{f_add_indicators}
\alias{f_add_indicators}
\title{Replace indicators to coin}
\usage{
f_add_indicators(coin, add_indicators = NULL)
}
\arguments{
\item{coin}{The coin}

\item{add_indicators}{character vector of indicator codes to replace}
}
\value{
updated coin
}
\description{
As \code{\link[=f_remove_indicators]{f_remove_indicators()}} but adds indicators back in. Obviously only
indicators that were originally present in the input data can be added.
}
