% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{f_build_GTM_coin}
\alias{f_build_GTM_coin}
\title{Shortcut function for building GTM coin, for debugging}
\usage{
f_build_GTM_coin()
}
\value{
coin
}
\description{
Shortcut function for building GTM coin, for debugging
}
