% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{f_build_index}
\alias{f_build_index}
\title{Calculate index results}
\usage{
f_build_index(coin, agg_method = "a_amean", only_aggregate = FALSE)
}
\arguments{
\item{coin}{The coin}

\item{agg_method}{An aggregation method (function accessible by COINr)}

\item{only_aggregate}{Logical: if \code{TRUE} just applies the aggregation step,
to save calculation time. This is used if only the aggregation method is changed,
or for reweighting.}
}
\value{
Updated coin
}
\description{
Calculates results following data treatment, normalisation and aggregation.
}
\details{
This function is invoked when the user arrives at the "Results" tab. It is
also used to recalculate results. Additionally generates the "severity" (1-5)
scores.
}
