% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_profiles.R
\name{f_compare_units_df}
\alias{f_compare_units_df}
\title{Comparison table for two units}
\usage{
f_compare_units_df(coin, usel1, usel2)
}
\arguments{
\item{coin}{The coin}

\item{usel1}{A unit, referenced by its uCode}

\item{usel2}{A unit, referenced by its uCode}
}
\value{
A data frame
}
\description{
A data frame with scores, ranks, severity, for all levels.
}
