% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_profiles.R
\name{f_compare_units_table}
\alias{f_compare_units_table}
\title{Interactive unit comparison table}
\usage{
f_compare_units_table(coin, usel1, usel2, using)
}
\arguments{
\item{coin}{The coin}

\item{usel1}{A unit, referenced by its uCode}

\item{usel2}{A unit, referenced by its uCode}

\item{using}{One of \code{c("Ranks", "Scores", "Severity")}}
}
\value{
A DT table with highlighting
}
\description{
Calls \code{\link[=f_compare_units_df]{f_compare_units_df()}}, then just shows the columns specified by \code{using}.
This is in order not to have too much info in the table. Returns DT table.
}
