% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_input.R
\name{f_data_input}
\alias{f_data_input}
\title{Data input}
\usage{
f_data_input(file_path, df_geom, ISO3 = NULL, uCode_col = NULL)
}
\arguments{
\item{file_path}{path to the excel file where we have the raw data}

\item{df_geom}{sf-class geometry data frame (used to cross check input codes)}

\item{ISO3}{ISO3 code of country to which the data belongs, e.g. \code{"GTM"}}

\item{uCode_col}{Column name of \code{df_geom} which is used as unit codes.}
}
\value{
coin-class object
}
\description{
Reads a formatted Excel file found at \code{file_path} and outputs a constructed coin.
}
\details{
On reading the Excel file, this function does the following:
\itemize{
\item Data is split into data and metadata and tidied
\item Metadata is read from structure tab
\item Any indicators with no data at all are removed
\item Any indicators with only one unique value are also removed (these cannot be normalised as they
have a range of zero)
\item Any resulting aggregation groups with no "children" are removed
\item A coin is assembled using COINr and this is the function output
}

If indicators/groups are removed, a message is sent to the console.

The Excel file is required to be in a fairly strict format: templates are downloaded
through the app or using \code{\link[=f_generate_input_template]{f_generate_input_template()}}.

User errors are trapped as much as possible with hopefully-helpful error messages.
}
