% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_compare.R
\name{f_df_for_scat}
\alias{f_df_for_scat}
\title{Simple comparison table for scatter plot}
\usage{
f_df_for_scat(coin, l, comp_with)
}
\arguments{
\item{coin}{The coin}

\item{l}{List of scenarios to compare}

\item{comp_with}{Either "Ranks" or "Scores"}
}
\value{
A data frame
}
\description{
Either merged table of scores or ranks. This is independent from the
\code{f_make_comparison_table()} function to allow some flexibility.
}
