% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_profiles.R
\name{f_display_group_table}
\alias{f_display_group_table}
\title{Interactive table of group scores}
\usage{
f_display_group_table(coin, usel, plot_group)
}
\arguments{
\item{coin}{The coin}

\item{usel}{Selected unit (uCode)}

\item{plot_group}{iCode of group to display}
}
\value{
A DT table
}
\description{
This is the counterpart table to the radar plot. Returns a table of scores or
values for a given aggregation group, including ranks and medians.
}
