% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_analysis.R
\name{f_display_indicator_analysis}
\alias{f_display_indicator_analysis}
\title{Extract and display analysis table}
\usage{
f_display_indicator_analysis(coin, filter_to_flagged = TRUE)
}
\arguments{
\item{coin}{The coin, with analysis tables in}

\item{filter_to_flagged}{Only show indicators with at least on flag.}
}
\value{
DT table
}
\description{
Extract and display analysis table
}
