% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{f_display_results_table}
\alias{f_display_results_table}
\title{Interactive results table}
\usage{
f_display_results_table(coin, type = "scores", as_discrete = FALSE)
}
\arguments{
\item{coin}{The coin}

\item{type}{Either "scores" or "ranks".}

\item{as_discrete}{If \code{TRUE}, displays the "severity" scores.}
}
\value{
A DT table
}
\description{
Returns a DT table for displaying results. Can toggle between scoes and ranks,
and displaying "severity" (1-5).
}
