% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{f_dset_to_severity}
\alias{f_dset_to_severity}
\title{Convert data set to severity scale}
\usage{
f_dset_to_severity(coin, iData)
}
\arguments{
\item{coin}{The coin}

\item{iData}{iData-format data frame}
}
\value{
Modified iData data frame, converted to severity categories
}
\description{
Converts an iData-formatted data set to the 1-5 severity scale.
Note this is only applied to levels > 1, i.e. not the indicator level.
}
