% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_export.R
\name{f_export_to_excel}
\alias{f_export_to_excel}
\title{Export to excel}
\usage{
f_export_to_excel(coin, l_scen, fname = "index_export.xlsx")
}
\arguments{
\item{coin}{The coin, with results present.}

\item{l_scen}{List of scenario data frames}

\item{fname}{file name to write the results to.}
}
\value{
Writes an Excel spreadsheet.
}
\description{
Simplified export to Excel with some formatting
Note if we change the index structure, the conditional formatting will need
to be adjusted.
}
