% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_analysis.R
\name{f_gather_correlations}
\alias{f_gather_correlations}
\title{Formatting for correlations}
\usage{
f_gather_correlations(X)
}
\arguments{
\item{X}{output from COINr::get_corr_flags()}
}
\value{
A vector of correlations
}
\description{
Formatting for correlations
}
