% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_input.R
\name{f_generate_input_template}
\alias{f_generate_input_template}
\title{Generate Excel template for input data, for specified country}
\usage{
f_generate_input_template(
  df_geom = NULL,
  to_file_name = NULL,
  uCode_col = NULL,
  uName_col = NULL,
  with_fake_data = FALSE
)
}
\arguments{
\item{df_geom}{Data frame of "sf" class which contains geometry for mapping the data,
and associated codes and names of each geographical area (region/country).}

\item{to_file_name}{Where to write the template file to.}

\item{uCode_col}{Column name in \code{df_geom} to use as uCodes in COINr. Must contain
unique codes, not starting with number.}

\item{uName_col}{Column name in \code{df_geom} to use as uNames in COINr (could also
be same as \code{uCode_col}).}

\item{with_fake_data}{Logical: if \code{TRUE} the input template will also contain
randomly-generated input data which can be used as a quick demo and for testing
uploads of new shape files.}
}
\description{
Writes codes and names of Admin2 regions to an Excel template, for a specified
country. Takes an sf-class data frame and extracts codes and names from it.
}
