% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{f_generate_results}
\alias{f_generate_results}
\title{Add results tables to coin}
\usage{
f_generate_results(coin)
}
\arguments{
\item{coin}{Coin}
}
\value{
Updated coin
}
\description{
Adds both a scores and ranks results table to the coin.
}
