% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{f_get_equal_weights}
\alias{f_get_equal_weights}
\title{Data frame of equal weights}
\usage{
f_get_equal_weights(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
Data frame of weights
}
\description{
Copies the original weights in the coin and sets them all equal
}
