% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{f_get_last_weights}
\alias{f_get_last_weights}
\title{Find last weights used}
\usage{
f_get_last_weights(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
Data frame of weights.
}
\description{
Returns the last set of weights used when aggregating the index, by checking
the log. Or \code{NULL} if not aggregated yet.
}
