% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{f_get_scenarios}
\alias{f_get_scenarios}
\title{Get results tables for all scenarios}
\usage{
f_get_scenarios(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
A list of data frames
}
\description{
Returns a list of results tables for all scenarios (aggregation methods).
Note that the aggregation methods are hard-coded here: will have to be updated
if we add new methods.
}
