% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_analysis.R
\name{f_highlight_DT}
\alias{f_highlight_DT}
\title{Generate display table for indicator analysis}
\usage{
f_highlight_DT(Xd, Xh, table_caption = NULL, highlight_colour = "#FAEB00")
}
\arguments{
\item{Xd}{Analysis table from \code{\link[=f_analyse_indicators]{f_analyse_indicators()}}}

\item{Xh}{Highlight table, also from \code{\link[=f_analyse_indicators]{f_analyse_indicators()}}}

\item{table_caption}{Optional caption}

\item{highlight_colour}{Highlight colour}
}
\value{
DT table
}
\description{
Takes analysis table and corresponding flags, and generates DT table with
highlighted cells.
}
