% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_profiles.R
\name{f_indicator_df}
\alias{f_indicator_df}
\title{Data frame of score, ranks etc for specific unit}
\usage{
f_indicator_df(coin, usel)
}
\arguments{
\item{coin}{The coin}

\item{usel}{A uCode found within the coin}
}
\value{
Data frame
}
\description{
This is like \code{\link[COINr:get_unit_summary]{COINr::get_unit_summary()}} but with some additions.
}
