% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_profiles.R
\name{f_indicator_table}
\alias{f_indicator_table}
\title{Interactive indicator table for a selected unit}
\usage{
f_indicator_table(coin, usel)
}
\arguments{
\item{coin}{The coin}

\item{usel}{A uCode found within the coin}
}
\value{
A DT table
}
\description{
A DT table for a selected unit
}
