% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_compare.R
\name{f_make_comparison_table}
\alias{f_make_comparison_table}
\title{Compare scenarios}
\usage{
f_make_comparison_table(coin, l, base_scen, comp_with, tab_type)
}
\arguments{
\item{coin}{The coin}

\item{l}{List of scenarios to compare}

\item{base_scen}{Name of scenario to use as the base scenario (for rank diffs)}

\item{comp_with}{Either "Ranks" or "Scores"}

\item{tab_type}{Either "Values", "Differences" or "Absolute differences"}
}
\value{
A data frame
}
\description{
Compares a list of scenarios as a data frame. The list \code{l} is a named list which is generated
by the Results tab, and includes a data frame of results for each tab.
}
