% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{f_make_severity_level_dset}
\alias{f_make_severity_level_dset}
\title{Make severity data set}
\usage{
f_make_severity_level_dset(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
Updated coin
}
\description{
Calculates the severity scores (1-5 scale) based on the aggregated data set,
and adds as a new data set called "Severity". Then uses this new data set
to calculate an additional results table.
}
