% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_profiles.R
\name{f_plot_radar}
\alias{f_plot_radar}
\title{Radar plot for a group}
\usage{
f_plot_radar(coin, usel, plot_group)
}
\arguments{
\item{coin}{The coin}

\item{usel}{Selected unit (uCode)}

\item{plot_group}{iCode of group to plot}
}
\value{
Radar plot, using plotly
}
\description{
Thin wrapper for \code{\link[iCOINr:iplot_radar]{iCOINr::iplot_radar()}}: just makes sure to return \code{NULL} if
less than 3 indicators in group.
}
