% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_analysis.R
\name{f_remove_indicators}
\alias{f_remove_indicators}
\title{Remove indicators from coin}
\usage{
f_remove_indicators(coin, remove_indicators = NULL)
}
\arguments{
\item{coin}{The coin}

\item{remove_indicators}{character vector of indicator codes to remove}
}
\value{
updated coin
}
\description{
Removes indicators as specified by a character vector of indicator
codes in \code{remove_indicators}. After removal, any results are regenerated
(updated). The original data is always preserved so indicators can be restored.
}
