% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{f_save_map}
\alias{f_save_map}
\title{Save map as image}
\usage{
f_save_map(plt, file_name = "map.png")
}
\arguments{
\item{plt}{Leaflet map object}

\item{file_name}{File path to save to, with file extension. The extension must
be one of .png, .pdf, .jpeg or .html}
}
\value{
File saved at specified path
}
\description{
First saves widget as html file using htmlwidgets. Then uses the webshot
package to save it as image.
}
