% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_compare.R
\name{f_style_comparison_table}
\alias{f_style_comparison_table}
\title{Style comparison table}
\usage{
f_style_comparison_table(df_merged, comp_with)
}
\arguments{
\item{df_merged}{Output of \code{\link[=f_make_comparison_table]{f_make_comparison_table()}}.}

\item{comp_with}{Either "Ranks" or "Scores" (colours are reversed if Ranks)}
}
\value{
DT table with styling.
}
\description{
Returns DT table with styling.
}
