% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_analysis.R
\name{filter_to_flagged}
\alias{filter_to_flagged}
\title{Filter analysis table to flagged}
\usage{
filter_to_flagged(Xd, Xh)
}
\arguments{
\item{Xd}{Stats table}

\item{Xh}{Flag table (for highlighting)}
}
\value{
List of data frames.
}
\description{
Only show rows with at least one flag - filters both tables.
}
