% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_profiles.R
\name{format_sw}
\alias{format_sw}
\title{Formatter for strengths and weaknesses}
\usage{
format_sw(X)
}
\arguments{
\item{X}{Output from \code{\link[COINr:get_str_weak]{COINr::get_str_weak()}}}
}
\value{
Tidied table
}
\description{
Tidies up formatting of numbers
}
