% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{get_aggregation_name}
\alias{get_aggregation_name}
\title{Get aggregation method names}
\usage{
get_aggregation_name(f_ag)
}
\arguments{
\item{f_ag}{Function name of aggregation method.}
}
\value{
String
}
\description{
Helper translating from COINr function names to full names.
}
