% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{get_cached_countries}
\alias{get_cached_countries}
\title{Get countries with cached maps}
\usage{
get_cached_countries()
}
\value{
Character vector of ISO3 codes
}
\description{
Returns ISO3 codes of all countries for which simplified maps have been
stored in inst/geom. Countries not in this list will have to have their
maps downloaded from the server.
}
