% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{get_codes_at_level}
\alias{get_codes_at_level}
\title{Get iCodes at specified level}
\usage{
get_codes_at_level(coin, Level)
}
\arguments{
\item{coin}{The coin}

\item{Level}{Index level}
}
\value{
Character vector of iCodes
}
\description{
Returns all iCodes at a specified level
}
