% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{get_index_code}
\alias{get_index_code}
\title{Get the highest level iCode}
\usage{
get_index_code(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
An iCode (string)
}
\description{
Returns the iCode at the highest level, i.e. the index code.
}
