% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{get_index_score}
\alias{get_index_score}
\title{Get score of selected unit at index level}
\usage{
get_index_score(coin, usel)
}
\arguments{
\item{coin}{The coin}

\item{usel}{A unit (uCode)}
}
\value{
A score
}
\description{
Get score of selected unit at index level
}
