% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_indicator_codes}
\alias{get_indicator_codes}
\title{Get list of available indicator/aggregate codes}
\usage{
get_indicator_codes(
  coin,
  code_types = "all",
  with_levels = TRUE,
  use_names = TRUE
)
}
\arguments{
\item{coin}{The coin}

\item{code_types}{Either \code{"Aggregate"} (for just aggregate codes), \code{"Indicator"} (for
just indicator codes) or \code{"all"} for both.}

\item{with_levels}{If \code{TRUE}, outputs a named list where the list names are of the format
"Level: iCode" or "Level: iName". These names are what users will see in the dropdown menus.
Else if \code{FALSE} just outputs a character vector of iCodes.}

\item{use_names}{if \code{TRUE} uses iNames in the named list, else uses iCodes.}
}
\value{
A list or character vector, for input into e.g. \code{selectInput()}.
}
\description{
This is for use in dropdown menus, e.g. in plotting.
}
