% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{get_indicator_direction}
\alias{get_indicator_direction}
\title{Get direction of iCode}
\usage{
get_indicator_direction(coin, iCode)
}
\arguments{
\item{coin}{The coin}

\item{iCode}{An iCode}
}
\value{
Either -1 or 1
}
\description{
Returns the direction of iCode
}
