% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_indicator_info}
\alias{get_indicator_info}
\title{HTML summary of a selected indicator}
\usage{
get_indicator_info(coin_full, coin, isel)
}
\arguments{
\item{coin_full}{The full coin with no indicators removed}

\item{coin}{The coin}

\item{isel}{A selected indicator (iCode)}
}
\value{
A text summary, as html.
}
\description{
Used in the analysis tab.
}
