% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{get_level_of_icode}
\alias{get_level_of_icode}
\title{Get level of iCode}
\usage{
get_level_of_icode(coin, iCode)
}
\arguments{
\item{coin}{The coin}

\item{iCode}{An iCode}
}
\value{
Level, as integer
}
\description{
Returns the level of iCode in the framework
}
