% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{get_n_indicators}
\alias{get_n_indicators}
\title{Get total number of indicators}
\usage{
get_n_indicators(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
Number
}
\description{
Get total number of indicators
}
