% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{get_number_of_children}
\alias{get_number_of_children}
\title{Find number of children}
\usage{
get_number_of_children(coin, iCode)
}
\arguments{
\item{coin}{The coin}

\item{iCode}{An indicator/aggregate}
}
\value{
Number
}
\description{
Returns number of children of specified indicator/aggregate
}
