% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{get_parent}
\alias{get_parent}
\title{Get parent at specified level}
\usage{
get_parent(coin, iCode, at_level = 2, as_name = TRUE)
}
\arguments{
\item{coin}{The coin}

\item{iCode}{iCode to check for parent}

\item{at_level}{Level to check for parent}

\item{as_name}{Logical: if \code{TRUE} returns iName}
}
\value{
iCode or iName of parent
}
\description{
Returns the parent group of an iCode at a specified level, either as iCode or
iName.
}
