% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{get_unit_summary_sev}
\alias{get_unit_summary_sev}
\title{Unit summary using severity level}
\usage{
get_unit_summary_sev(coin, usel, Level)
}
\arguments{
\item{coin}{The coin}

\item{usel}{Selected unit (uCode)}

\item{Level}{Level at which to return scores}
}
\value{
A data frame
}
\description{
Returns a unit summary based on the severity level. For use in the map panel
This works like \code{\link[COINr:get_unit_summary]{COINr::get_unit_summary()}} but for the severity category.
}
