% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GTM_coin}
\alias{GTM_coin}
\title{Example coin for use in tests}
\format{
A "coin" object
}
\source{
Data from UNCHR Guatemala
}
\usage{
GTM_coin
}
\description{
This is a coin assembled from example data from Guatemala. It is only used
for the purposes of fast testing.
}
\keyword{datasets}
