% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{header_help_icon}
\alias{header_help_icon}
\title{Help pop up window}
\usage{
header_help_icon(id_modal)
}
\arguments{
\item{id_modal}{ID of modal to display}
}
\value{
HTML
}
\description{
Function that creates a help pop up window for each page. This function
generates the icon, the thing that is displayed is controlled by \code{id_modal}.
This is called in the app server function.
}
