% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_COINr.R
\name{reset_coin}
\alias{reset_coin}
\title{Replace all indicators}
\usage{
reset_coin(coin)
}
\arguments{
\item{coin}{The coin}
}
\value{
Updated coin
}
\description{
Reset coin so that all indicators are included, and regenerate results.
Note that any analysis tables etc will be lost.
}
