% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_input.R
\name{validate_iMeta}
\alias{validate_iMeta}
\title{Verbose check on indicator metadata tab}
\usage{
validate_iMeta(iMeta)
}
\arguments{
\item{iMeta}{Table imported from user input spreadsheet.}
}
\value{
Message as text, or \code{NULL} if no errors.
}
\description{
Checks and messages to pass to the user for basic errors in metadata rows of
the Data tab in the input template. Called in \code{\link[=f_data_input]{f_data_input()}}.
}
