% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{weights_slider}
\alias{weights_slider}
\title{Weights slider}
\usage{
weights_slider(id, label = NULL, value = 1)
}
\arguments{
\item{id}{ID for slider}

\item{label}{Label}

\item{value}{initial value}
}
\value{
HTML code for slider
}
\description{
Thin wrapper for sliderinput.
}
