% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_input.R
\name{f_generate_input_template}
\alias{f_generate_input_template}
\title{Generate Excel template for input data, for specified country}
\usage{
f_generate_input_template(ISO3, to_file_name = NULL)
}
\arguments{
\item{ISO3}{Valid ISO3 code with geometry available in inst/geom}

\item{to_file_name}{Where to write to}
}
\description{
Writes codes and names of Admin2 regions to an Excel template, for a specified
country.
}
