% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_results.R
\name{f_plot_map}
\alias{f_plot_map}
\title{Choropleth map of indicator/aggregate}
\usage{
f_plot_map(
  coin,
  iCode,
  ISO3,
  as_discrete = TRUE,
  bin_colours = NULL,
  poly_opacity = 0.7,
  line_colour = "white",
  line_weight = 2,
  line_type = "3",
  legendposition = "bottomright"
)
}
\arguments{
\item{coin}{The coin}

\item{iCode}{iCode of indicator/aggregate to plot}

\item{ISO3}{Country of the data to plot}

\item{as_discrete}{If \code{TRUE}, plots severity categories.}

\item{bin_colours}{Vector of colours to use on discrete palette if \code{as_discrete = TRUE}}

\item{poly_opacity}{Opacity for polygons: value between 0 and 1}

\item{line_colour}{Colour for lines}

\item{line_weight}{Weight for lines}

\item{line_type}{Type for lines: value from 1-4.}

\item{legendposition}{Legend position argument passed to Leaflet}
}
\value{
Leaflet map object
}
\description{
Uses leaflet and stored geometry files to plot any selected indicator or
aggregate as a choropleth map, for a selected country.
}
\details{
This function relies on geometry files that are stored inside the A2SIT package.
It will therefore not work if the country selected by \code{ISO3} does not have
available geometry.

To see which countries currently have available geometry, run: \code{\link[=get_cached_countries]{get_cached_countries()}}.

The function merges the selected indicator with the geometry file, using the
uCodes. At this point in the app, it should have been checked that the user's
uCodes correspond exactly with those in the geometry file.

To change colours and styling, look in the source code of this function.
}
